create a trello card in addition to email for escalation
Do not rely on rel-del tool to scan for email and open a card.
This will prevent duplicate emails from being sent when other
scripts fail or are not running.

Include the key "proa" in the subject of the escalation to
indicate to other scripts not to create a card.